requirements.txt
customtkinter==5.2.2
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
